pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "code-review-agent"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "python-dotenv>=1.0.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "gitpython>=3.1.0",
]

[project.optional-dependencies]
gemini = ["google-genai>=1.0.0"]
groq   = ["groq>=0.9.0"]
ui     = ["streamlit>=1.35.0"]

[tool.setuptools.packages.find]
where = ["."]
include = ["agent*"]
